% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_irr_summary}
\alias{all_irr_summary}
\title{Irrigation Volume by Month and Year}
\format{

}
\usage{
all_irr_summary
}
\description{
This table summarizes the total estimated irrigation volume (using OpenET; source: https://www.dri.edu/project/owrd-et/) for all fields impacted by potential maximum inundation. A threshold of 30% was used to determine which fields were impacted.
}
\keyword{datasets}
