% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodplain_acres}
\alias{floodplain_acres}
\title{Estimated Maximum Floodplain Acres}
\format{

}
\usage{
floodplain_acres
}
\description{
This data object summarizes the theoretical maximum inundation for chosen reaches in the Klamath basin as well as the potential agricultural impacts
}
\keyword{datasets}
